Add Anaconda UI version to BZ report
Dynamically add the file VERSION.txt which contains the exact version
used on the specfile. Ideally the version could be detected by simply
parsing the output of `rpm -q anaconda-webui`, but that cannot be
detected on the VM used for development.

This approach has limitations, though. The file is destroyed when
running `make rsync` as is.

Relates: INSTALLER-4270